requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
configparser==4.0.2
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
redis==3.4.1
requests==2.23.0
schedule==0.6.0
urllib3==1.25.8
Werkzeug==1.0.0
python-crontab
